refactor: 인증(Authentication) 모듈 폴더 구조 개선 및 역할별 분리 #9
