+++ src/main/java/com/parkez/auth/authentication/refresh/RefreshToken.java
@@ -1,4 +1,4 @@
-package com.parkez.auth.refresh;
+package com.parkez.auth.authentication.refresh;
 
 import com.parkez.common.entity.BaseEntity;
