+++ src/main/java/com/parkez/auth/authentication/refresh/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.parkez.auth.refresh;
+package com.parkez.auth.authentication.refresh;
 
 import org.springframework.data.jpa.repository.JpaRepository;
